Simple Moving Average functionality with test
