+++ components/express-traderr/src/logic/sma.js
