+++ index.html
@@ -26,8 +26,10 @@
     <title>express-traderr</title>
   </head>
   <body>
+    <script src="/__/firebase/7.8.0/firebase-app.js"></script>
     <express-traderr title="Hello owc world!"></express-traderr>
     <script type="module" src="./components/express-traderr/express-traderr.js"></script>
   </body>
-  <script type="module" src="./components/express-traderr/src/firebase.js" ></script>
+
+  <script type="module" src="./components/express-traderr/src/firebase.js"></script>
 </html>
